Renaming ViewSampleDialog to ConfigurationSamplePreviewDialog
