+++ ui/app/src/components/ConfigurationSamplePreviewDialog/styles.ts
